+++ __SITE__/analyse/home/helpers.rb
@@ -9,17 +9,17 @@ class Analyse
     #               film. Ça peut être un simple visiteur, un inscrit,
     #               un analyste, etc.
     def list_for reader
-      where =
-        case 
-        when !reader.identified? then '100'
-        when reader.suscriber?  then '1'
-        else '11'
-        end
-      where = "WHERE SUBSTRING(fa.specs,1,#{where.length}) = '#{where}'"
-      debug "where = #{where}"
+
+
+      # On ne met plus de filtre sur le reader, maintenant : tout le
+      # monde peut voir les films à partir du moment où ils sont 
+      # analysés.
+      where = "WHERE SUBSTRING(fa.specs,1,1) = '1'"
+      
       lis_films = String.new
       request = <<-SQL
-      SELECT f.id, f.titre, f.titre_fr, fa.specs
+      SELECT f.id, f.titre, f.titre_fr, fa.specs, f.annee,
+        fa.realisateur AS director
         FROM films_analyses fa
         INNER JOIN filmodico f ON fa.id = f.id
         #{where};
@@ -27,9 +27,27 @@ class Analyse
       site.db.use_database(:biblio)
       site.db.execute(request)
         .each do |hfilm|
-        lis_films << "<li class=\"film\" id=\"film-#{hfilm[:id]}\">#{hfilm[:titre]}</li>"
+        lis_films << li_film(hfilm)
       end
       return "<ul id=\"analyse_list\">#{lis_films}</ul>"
+    end
+    
+    def li_film hfilm
+      <<-HTML
+      <li class="film" id="film-#{hfilm[:id]}">
+        <span class="titre">
+          <a href="analyse/lire/#{hfilm[:id]}" class="nodeco">
+            #{hfilm[:titre].force_encoding('utf-8')}
+          </a>
+        </span>
+        <span class="annee">#{hfilm[:annee]}</span>
+        <span class="director">#{only_name_of hfilm[:director]}</span>
+      </li>
+      HTML
+    end
+
+    def only_name_of patronyme
+      patronyme.force_encoding('utf-8').split(' ').last
     end
   end #<< self
 end #/Analyse