+++ spec/support/analyse/main.rb
@@ -1,7 +1,9 @@
 # @param {Hash} params
 #               :access     :public, ou :inscrit ou :suscribed
-#                           Pour définir le niveau d'accès des films à 
+#                           Pour définir le niveau d'accès des films à
 #                           prendre. Correspond à l'accès du visiteur.
+#               :all        Si true, retourne tous les films correspondants
+#                           à la rechercher
 def get_film_analyse params = nil
   defined?(Site) || require_lib_site
   params ||= Hash.new
@@ -21,14 +23,26 @@ def get_film_analyse params = nil
       "WHERE #{where_clause.join(' AND ')}"
     else '' end
   request = <<-SQL
-  SELECT fa.*
+  SELECT fa.*, fa.realisateur AS director,
+    f.*
     FROM films_analyses fa
     INNER JOIN filmodico f ON f.id = fa.id
     #{where_clause}
-    LIMIT 1
   SQL
+  unless params[:all]
+    request << ' LIMIT 1'
+  end
   site.db.use_database(:biblio)
-  hfilm = site.db.execute(request).first
-  expect(hfilm).not_to eq nil
-  return hfilm
+  if params[:all]
+    return site.db.execute(request)
+  else
+    # Un seul film
+    hfilm = site.db.execute(request).first
+    expect(hfilm).not_to eq nil
+    return hfilm
+  end
+end
+
+def get_films_analyse params = nil
+  get_film_analyse params.merge(all: true)
 end